pyproject.toml
[tool.poetry]
name = "alchemy-lab"
version = "0.1.0"
description = ""
authors = ["jclarte <j.clarte@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.0.5"
sqlalchemy = "^2.0.20"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
